// src/api/types.gen.ts
export type CurriculumComplete = {
    /**
     * Name
     */
    name: string;
    /**
     * Id
     */
    id: string;
};
// ...
export type DocumentComplete = {
    /**
     * Name
     */
    name: string;
    /**
     * Id
     */
    id: string;
    /**
     * Seller Id
     */
    seller_id: string;
};
// ...
export type GenerateTicketComplete = {
    /**
     * Name
     */
    name: string;
    /**
     * Max Use
     */
    max_use: number;
    /**
     * Expiration
     */
    expiration: string;
    /**
     * Id
     */
    id: string;
};
// ...
export type MembershipSimple = {
    /**
     * Name
     */
    name: string;
    /**
     * Manager Group Id
     */
    manager_group_id: string;
    /**
     * Id
     */
    id: string;
};
// ...
export type ProductCompleteNoConstraint = {
    /**
     * Name Fr
     */
    name_fr: string;
    /**
     * Name En
     */
    name_en?: string | null;
    /**
     * Description Fr
     */
    description_fr?: string | null;
    /**
     * Description En
     */
    description_en?: string | null;
    /**
     * Available Online
     */
    available_online: boolean;
    /**
     * Needs Validation
     */
    needs_validation: boolean;
    /**
     * Year
     */
    year: number;
    /**
     * Id
     */
    id: string;
    /**
     * Seller Id
     */
    seller_id: string;
    /**
     * Variants
     */
    variants?: Array<AppModulesCdrSchemasCdrProductVariantComplete>;
    related_membership?: MembershipSimple | null;
    /**
     * Tickets
     */
    tickets: Array<GenerateTicketComplete>;
};
// ...
export type AppModulesCdrSchemasCdrProductComplete = {
    /**
     * Name Fr
     */
    name_fr: string;
    /**
     * Name En
     */
    name_en?: string | null;
    /**
     * Description Fr
     */
    description_fr?: string | null;
    /**
     * Description En
     */
    description_en?: string | null;
    /**
     * Available Online
     */
    available_online: boolean;
    /**
     * Needs Validation
     */
    needs_validation: boolean;
    /**
     * Id
     */
    id: string;
    /**
     * Year
     */
    year: number;
    /**
     * Seller Id
     */
    seller_id: string;
    /**
     * Variants
     */
    variants?: Array<AppModulesCdrSchemasCdrProductVariantComplete>;
    related_membership?: MembershipSimple | null;
    /**
     * Product Constraints
     */
    product_constraints?: Array<ProductCompleteNoConstraint>;
    /**
     * Document Constraints
     */
    document_constraints?: Array<DocumentComplete>;
    /**
     * Tickets
     */
    tickets?: Array<GenerateTicketComplete>;
};

/**
 * ProductEdit
 */
export type AppModulesCdrSchemasCdrProductEdit = {
    /**
     * Name Fr
     */
    name_fr?: string | null;
    /**
     * Name En
     */
    name_en?: string | null;
    /**
     * Description Fr
     */
    description_fr?: string | null;
    /**
     * Description En
     */
    description_en?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Available Online
     */
    available_online?: boolean | null;
    related_membership?: MembershipSimple | null;
    /**
     * Product Constraints
     */
    product_constraints?: Array<string> | null;
    /**
     * Document Constraints
     */
    document_constraints?: Array<string> | null;
};
// ...
export type AppModulesCdrSchemasCdrProductVariantComplete = {
    /**
     * Id
     */
    id: string;
    /**
     * Year
     */
    year: number;
    /**
     * Product Id
     */
    product_id: string;
    /**
     * Name Fr
     */
    name_fr: string;
    /**
     * Name En
     */
    name_en?: string | null;
    /**
     * Description Fr
     */
    description_fr?: string | null;
    /**
     * Description En
     */
    description_en?: string | null;
    /**
     * Price
     */
    price: number;
    /**
     * Enabled
     */
    enabled: boolean;
    /**
     * Unique
     */
    unique: boolean;
    /**
     * Allowed Curriculum
     */
    allowed_curriculum?: Array<CurriculumComplete>;
    /**
     * Related Membership Added Duration
     */
    related_membership_added_duration?: string | null;
};